+++ src/App.js
@@ -1,13 +1,13 @@
-import './App.css';
+import "./App.css";
+import { ProductScreen } from "./containers/products/ProductScreen";
 
 function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-       <h1>Shopping Cart</h1>
-      </header>
-    </div>
-  );
+	return (
+		<div className="App">
+			<h1>Shopping Cart</h1>
+			<ProductScreen />
+		</div>
+	);
 }
 
 export default App;
